@@ -5,7 +5,7 @@ requires = ["hatchling", "hatch-vcs"]
 [project]
 name = "anndata"
 description = "Annotated data."
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = "BSD-3-Clause"
 authors = [
     { name = "Philipp Angerer" },
@@ -29,7 +29,6 @@ classifiers = [
     "Operating System :: Microsoft :: Windows",
     "Operating System :: POSIX :: Linux",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -43,7 +42,7 @@ dependencies = [
     "numpy>=1.23",
     # https://github.com/scverse/anndata/issues/1434
     "scipy >1.8",
-    "h5py>=3.1",
+    "h5py>=3.7",
     "exceptiongroup; python_version<'3.11'",
     "natsort",
     "packaging>=20.0",
@@ -75,7 +74,7 @@ doc = [
     "nbsphinx",
     "scanpydoc[theme,typehints] >=0.14.1",
     "zarr",
-    "awkward>=2.0.7",
+    "awkward>=2.3",
     "IPython",                          # For syntax highlighting in notebooks
     "myst_parser",
     "sphinx_design>=0.5.0",
@@ -86,7 +85,7 @@ doc = [
 dev-doc = ["towncrier>=24.8.0"] # release notes tool
 test = [
     "loompy>=3.0.5",
-    "pytest>=8.2",
+    "pytest>=8.2,<8.3.4",
     "pytest-cov>=2.10",
     "zarr<3.0.0a0",
     "matplotlib",
@@ -94,7 +93,7 @@ test = [
     "openpyxl",
     "joblib",
     "boltons",
-    "scanpy",
+    "scanpy>=1.9.8",
     "httpx", # For data downloading
     "dask[distributed]",
     "awkward>=2.3",
@@ -108,7 +107,7 @@ gpu = ["cupy"]
 cu12 = ["cupy-cuda12x"]
 cu11 = ["cupy-cuda11x"]
 # https://github.com/dask/dask/issues/11290
-dask = ["dask[array]>=2022.09.2,<2024.8.0"]
+dask = ["dask[array]>=2022.09.2,!=2024.8.*,!=2024.9.*"]
 
 [tool.hatch.version]
 source = "vcs"
@@ -182,6 +181,7 @@ select = [
     "ICN", # Follow import conventions
     "PTH", # Pathlib instead of os.path
     "PT",  # Pytest conventions
+    "PYI", # Typing
 ]
 ignore = [
     # line too long -> we accept long comment lines; formatter gets rid of long code lines
